protected views from non admins and must be logged in to access cart

+++ app/controllers/suppliers_controller.rb
@@ -1,12 +1,24 @@
 class SuppliersController < ApplicationController
   def index
+    unless current_user && current_user.admin == 1
+      redirect_to "/"
+      return
+    end
     @suppliers = Supplier.all
     render = "index.html.erb"
   end
   def new
+    unless current_user && current_user.admin == 1
+      redirect_to "/"
+      return
+    end
     render = "new.html.erb"
   end
   def create
+    unless current_user && current_user.admin == 1
+      redirect_to "/"
+      return
+    end
     supplier = Supplier.new(name: params[:name], email: params[:email], phone_number: params[:phone_number])
     supplier.save
     flash[:success]= "Supplier successfully created."
@@ -19,6 +31,10 @@ class SuppliersController < ApplicationController
   def update
   end
   def destroy
+    unless current_user && current_user.admin == 1
+      redirect_to "/"
+      return
+    end
     supplier = Supplier.find_by(id: params[:id])
     supplier.destroy
     flash[:success] = "Supplier successfully destroyed."
